Fixed incorrect passed value for formatting of object address.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@35378 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/except.c
@@ -2284,7 +2284,7 @@ rt_private void print_object_location_reason_effect (
 
 		/* Print object address with 16 digits to be ready when pointers
 		 * will be on 64 bits on all platform. */
-	sprintf(buffer, "<%016lX>  ", a_object_addr);
+	sprintf(buffer, "<%016lX>  ", (unsigned long) a_object_addr);
 	append_trace(buffer);
 
 	if (l_location_count > 19) {
